+++ src/components/ui/index.js
@@ -9,28 +9,28 @@ import MyColorSelect from "@/components/ui/pinia_ui/MyColorSelect.vue";
 // PINIA
 
 
-import MySelection from "@/components/ui/MySelection.vue";
-import MyController from "@/components/ui/MyController.vue";
+import MySelection from "@/components/ui/pinia_ui/MySelection.vue";
+import MyController from "@/components/ui/pinia_ui/MyController.vue";
 import MySlider from "@/components/ui/MySlider.vue";
-import MyButton from "@/components/ui/MyButton.vue";
-import MyButtonBack from "@/components/ui/MyButtonBack.vue";
-import MyButtonGender from "@/components/ui/MyButtonGender.vue";
-import MyAttribute from "@/components/ui/MyAttribute.vue";
-import MyInventory from "@/components/ui/MyInventory.vue";
-import MyFines from "@/components/ui/MyFines.vue";
-import MySelectionBox from "@/components/ui/MySelectionBox.vue";
-import MySelectionCard from "@/components/ui/MySelectionCard.vue";
-import MyCardText from "@/components/ui/MyCardText.vue";
+import MyButton from "@/components/ui/pinia_ui/MyButton.vue";
+import MyButtonBack from "@/components/ui/pinia_ui/MyButtonBack.vue";
+import MyButtonGender from "@/components/ui/pinia_ui/MyButtonGender.vue";
+import MyAttribute from "@/components/ui/pinia_ui/MyAttribute.vue";
+import MyInventory from "@/components/ui/pinia_ui/MyInventory.vue";
+import MyFines from "@/components/ui/pinia_ui/MyFines.vue";
+import MySelectionBox from "@/components/ui/pinia_ui/MySelectionBox.vue";
+import MySelectionCard from "@/components/ui/pinia_ui/MySelectionCard.vue";
+import MyCardText from "@/components/ui/pinia_ui/MyCardText.vue";
 import MySpellText from "@/components/ui/MySpellText.vue";
-import MyWrapper from "@/components/ui/MyWrapper.vue";
+import MyWrapper from "@/components/ui/pinia_ui/MyWrapper.vue";
 import RaceBody from "@/components/ui/pinia_ui/RaceBody.vue";
 import MySkale from "@/components/ui/pinia_ui/MySkale.vue";
 import MySizeGrowth from "@/components/ui/pinia_ui/MySizeGrowth.vue";
 import MyRange from "@/components/ui/pinia_ui/MyRange.vue";
 import MyRangeSize from "@/components/ui/pinia_ui/MyRangeSize.vue";
-import MyRangeHight from "@/components/ui/MyRangeHight.vue";
-import MyHeaderCard from "@/components/ui/MyHeaderCard.vue";
-import MyLogoCard from "@/components/ui/MyLogoCard.vue";
+
+import MyHeaderCard from "@/components/ui/pinia_ui/MyHeaderCard.vue";
+import MyLogoCard from "@/components/ui/pinia_ui/MyLogoCard.vue";
 import MagicAttribute from "@/components/ui/MagicAttribute.vue";
 import MyDialogSpell from "@/components/ui/MyDialogSpell.vue";
 import MySpellMain from "@/components/ui/MySpellMain.vue";
@@ -66,7 +66,7 @@ export default [
   MySizeGrowth,
   MyRange,
   MyRangeSize,
-  MyRangeHight,
+
   MyHeaderCard,
   MyLogoCard,
   MagicAttribute,
