+++ src/components/ui/pinia_ui/MySelectionBox.vue
@@ -1,13 +1,4 @@
 <template>
-  <!-- <div
-    class="sidebar_wrap"
-    :class="{
-      sidebar_selection: Sid_Sel && !Sid_Sel_Stay,
-      sidebar_selection_open: shown && Sid_Sel && !Sid_Sel_Stay,
-      sidebar_selection_stayed: Sid_Sel_Stay,
-      sidebar_selection_stayed_open: shown && Sid_Sel_Stay,
-      }"
-  > -->
   <div
   class="sidebar_selection"
     :class="{
@@ -25,43 +16,16 @@
 <script>
 export default {
   name: "MySelectionBox",
-  data() {
-    return {
-      // home_arr: this.$root.main_page.shown_home_arr,
-    };
-  },
   props: {
     shown: {
       type: Boolean,
       default: false,
     },
   },
-  computed: {
-    // Sid_Sel() {
-    //   console.log('Sid_Sel', !this.home_arr[0] && this.home_arr[1])
-    //   return !this.home_arr[0] && this.home_arr[1]
-    // },
-    // Sid_Sel_Stay() {
-    //   console.log('Sid_Sel_Stay', this.home_arr[0] && this.home_arr[1])
-    //   return this.home_arr[0] && this.home_arr[1]
-    // },
-  },
-  methods: {
-  }
 };
 </script>
 
 <style scoped>
-  /* .sidebar_wrapers {
-  height: 100%;
-  position: absolute;
-  overflow-y: scroll;
-}
-
-.sidebar_wrapers::-webkit-scrollbar {
-  width: 0;
-} */
-
   .sidebar_selection {
   height: 100%;
   left: 426px;
@@ -80,18 +44,6 @@ export default {
   width: 0;
   display: none;
 }
-
-/* .sidebar_selection_stayed {
-  height: 100%;
-  position: absolute;
-  left: 426px;
-  opacity: 0;
-  overflow-y: scroll;
-  -webkit-transition: 0.4s ease-in-out;
-  -moz-transition: 0.4s ease-in-out;
-  -o-transition: 0.4s ease-in-out;
-  transition: 0.4s ease-in-out;
-} */
 
 .sidebar_selection_open {
   opacity: 1;
